+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Fix access to `let_it_be` variables in `after(:all)` hook. ([@cbarton][])
 
+- Add support for using the before_all hook with Rails' parallelize feature (using processes). ([@peret][])
+
+Make sure to include `TestProf::BeforeAll::Minitest` before you call `parallelize`.
+
 ## 1.0.6 (2021-06-23)
 
 - Fix Spring detection when `DISABLE_SPRING=1` is used. ([@palkan][])